create style sheet for navbar and refactor
