+++ src/App.js
@@ -7,7 +7,7 @@ import NotFound from "./components/NotFound";
 import AddItem from "./components/AddItem";
 import ListContainer from "./components/ShoppingList/ListContainer";
 import Home from "./components/Home";
-import Navbar from "./components/Navbar";
+import Navbar from "./components/Navbar/Navbar";
 
 const App = () => {
   const [auth, setAuth] = useState(localStorage.getItem("token"));
