+++ src/components/Navbar/Navbar.js
@@ -1,10 +1,12 @@
 import React from "react";
+import "./Navbar.scss";
+import "../../styles/base.scss";
 
-import { ReactComponent as MCB } from "../img/meat_cheese_bread.svg";
-import { ReactComponent as ListView } from "../img/list_view.svg";
-import { ReactComponent as AddItem } from "../img/add_item.svg";
-import { ReactComponent as ChangeList } from "../img/change_list.svg";
-import { ReactComponent as ShareList } from "../img/share_list.svg";
+import { ReactComponent as MCB } from "../../img/meat_cheese_bread.svg";
+import { ReactComponent as ListView } from "../../img/list_view.svg";
+import { ReactComponent as AddItem } from "../../img/add_item.svg";
+import { ReactComponent as ChangeList } from "../../img/change_list.svg";
+import { ReactComponent as ShareList } from "../../img/share_list.svg";
 
 import { Link } from "react-router-dom";
