+++ src/components/Navbar/Navbar.scss
@@ -1,0 +1,22 @@
+.navLink {
+  margin-left: 1em;
+}
+
+.homeLink {
+  display: flex;
+  align-items: center;
+  color: black;
+  margin-right: 2.5em;
+}
+
+.navContainer {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.navStyle {
+  display: flex;
+  align-items: center;
+  align-content: space-between;
+}
